chore: 컴파일러 및 Gradle 경고 제거
- FcmPushClient: @Value 어노테이션에 param use-site target 명시
- 불필요한 non-null assertion(!!) 제거 (LegacyCourseMetadata.id, TimetableBriefDto.id)
- Long 타입에 대한 중복 변환 호출(toLong()) 제거
- build.gradle.kts: deprecated 된 by tasks 위임 대신 tasks.getByName<BootJar>("bootJar") 사용

+++ v1compat/src/main/kotlin/com/wafflestudio/snutt/v1compat/snutt/V1CompatThemeDiaryControllers.kt
@@ -406,7 +406,7 @@ class V1CompatDiaryController(
             diaryService.generateQuestionnaire(
                 user.id!!,
                 DiaryQuestionnaireRequest(
-                    lectureId = body.lectureId.toLong(),
+                    lectureId = body.lectureId,
                     dailyClassTypes = body.dailyClassTypes,
                 ),
             )
@@ -488,7 +488,7 @@ class V1CompatDiaryController(
         diaryService.submitDiary(
             user.id!!,
             DiarySubmissionRequest(
-                lectureId = body.lectureId.toLong(),
+                lectureId = body.lectureId,
                 dailyClassTypes = body.dailyClassTypes,
                 questionAnswers = body.questionAnswers,
                 comment = body.comment,
